+++ lib/shadcn_ui.ex
@@ -21,6 +21,7 @@ defmodule ShadcnUI do
     ShadcnUI.Components.Navigation.SectionHeader,
     ShadcnUI.Components.Overlays.AlertDialog,
     ShadcnUI.Components.Overlays.Dialog,
+    ShadcnUI.Components.Overlays.Drawer,
     ShadcnUI.Components.Foundation.Alert,
     ShadcnUI.Components.Foundation.Avatar,
     ShadcnUI.Components.Foundation.Badge,